api/requirements.txt
fastapi==0.115.6
rembg==2.0.69
pillow==11.1.0
requests==2.32.3
python-multipart==0.0.20
onnxruntime==1.21.0
mangum==0.18.2
